Updating Some Things
The project is mostly complete (All core functionality is there) and
working correctly. I am making a few enhancements that are independent
of the core components of the experiment (So I won't break it by adding
these things). Going to start gathering data. Everything runs much
faster than I expected so that will not take long. 

I am also working on created a real time example. It is already almost
complete. As of now the text in the live example will try to correct the
entire string each time the phrase is checked (leading to some
characters that have been already corrected to change to another
possible candidate). Plan on fixing this by only passing new characters
to the correction functions.